replace original NT files
