--- Assets/Mirror/Components/Experimental/Oumuamua.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 08e0fc21b95a46ddbdef2dd29f16baae
-timeCreated: 1615775769
