--- Assets/Mirror/Components/Experimental/Oumuamua/OumuamuaBase.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6019c95737204054bc9539ddfb873d5d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: 7453abfe9e8b2c04a8a47eb536fe21eb, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
