--- Assets/Mirror/Components/Experimental/Oumuamua/OumuamuaChild.cs
@@ -1,13 +1,0 @@
-// A component to synchronize the position of child transforms of networked objects.
-// There must be a NetworkTransform on the root object of the hierarchy. There can be multiple NetworkTransformChild components on an object. This does not use physics for synchronization, it simply synchronizes the localPosition and localRotation of the child transform and lerps towards the recieved values.
-using UnityEngine;
-
-namespace Mirror.Experimental
-{
-    public class OumuamuaChild : OumuamuaBase
-    {
-        [Header("Target")]
-        public Transform target;
-        protected override Transform targetComponent => target;
-    }
-}
