--- Assets/Mirror/Tests/Editor/NetworkTransformTest.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3114c5bb742fe41c09ec88e90ac29ef4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
